+++ frontend/src/index.js
@@ -3,10 +3,10 @@ import ReactDOM from 'react-dom';
 
 import 'bootstrap/dist/css/bootstrap.min.css';
 
-import StudentView from './prompted_readings/student_view';
-import InstructorView from './prompted_readings/instructor_view';
-import {AnalysisView} from './prompted_readings/analysis_view';
-import ReadingView from './prompted_readings/reading_view';
+import StudentView from './views/student_view';
+import InstructorView from './views/instructor_view';
+import {AnalysisView} from './views/analysis_view';
+import ReadingView from './views/reading_view';
 
 
 window.app_modules = {
