+++ frontend/src/views/analysis_view.js
