+++ frontend/src/views/student_view.css
