+++ frontend/src/views/student_view.js
